+++ zcash_client_backend/src/lib.rs
@@ -64,6 +64,7 @@
 pub use zcash_keys::address;
 pub mod data_api;
 mod decrypt;
+use zcash_keys::address::Address;
 pub use zcash_keys::encoding;
 pub mod fees;
 pub use zcash_keys::keys;
@@ -107,6 +108,27 @@ pub enum PoolType {
     Transparent,
     /// A shielded value pool.
     Shielded(ShieldedProtocol),
+}
+
+impl PoolType {
+    pub fn is_receiver(&self, addr: &Address) -> bool {
+        match addr {
+            Address::Sapling(_) => matches!(self, PoolType::Shielded(ShieldedProtocol::Sapling)),
+            Address::Transparent(_) => matches!(self, PoolType::Transparent),
+            Address::Unified(ua) => match self {
+                PoolType::Transparent => ua.transparent().is_some(),
+                PoolType::Shielded(ShieldedProtocol::Sapling) => ua.sapling().is_some(),
+                #[cfg(zcash_unstable = "orchard")]
+                PoolType::Shielded(ShieldedProtocol::Orchard) => {
+                    #[cfg(feature = "orchard")]
+                    return ua.orchard().is_some();
+
+                    #[cfg(not(feature = "orchard"))]
+                    return false;
+                }
+            },
+        }
+    }
 }
 
 impl fmt::Display for PoolType {